pyproject.toml
[project]
name = "post-creater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "bs4>=0.0.2",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "distro==1.9.0",
    "frozenlist==1.7.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.10.0",
    "lxml==6.0.0",
    "multidict==6.6.3",
    "numpy==2.3.1",
    "openai==1.108.1",
    "packaging==25.0",
    "pandas==2.3.1",
    "primp==0.15.0",
    "propcache==0.3.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "readability-lxml>=0.8.4.1",
    "requests==2.32.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "tenacity>=9.1.2",
    "tqdm==4.67.1",
    "trafilatura>=2.0.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "yarl==1.20.1",
]

[dependency-groups]
dev = [
    "nuitka>=2.8.4",
]
[tool.basedpyright]
typeCheckingMode = "standard"
